Meters kunnen nu aangepast worden
-Meters kunnen aangepast worden
- creation_timestamp in database verandert naar biginteger

+++ EnerjoyApi/database/migrations/2020_02_26_162343_create_meters_table.php
@@ -16,7 +16,7 @@ class CreateMetersTable extends Migration
         Schema::create('meters', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('meter_id');
-            $table->integer('creation_timestamp')->nullable();
+            $table->bigInteger('creation_timestamp')->nullable();
             $table->tinyInteger('isUsed')->default(0);
             $table->tinyInteger('deleted')->default(0);
         });
